making one facet component and making scroll a utility function
